requirements.txt
Flask
requests
python-dotenv
Flask-SQLAlchemy
